+++ mods/BrewLAN_Gameplay/ParagonGame/units/SAB0104/SAB0104_script.lua
@@ -10,10 +10,10 @@
 local CConstructionStructureUnit = import('/lua/cybranunits.lua').CConstructionStructureUnit
 local EffectUtil = import('/lua/EffectUtilities.lua')
 
-PAB0104 = Class(CConstructionStructureUnit) 
+SAB0104 = Class(CConstructionStructureUnit)
 {
     CreateBuildEffects = function( self, unitBeingBuilt, order )
         EffectUtil.CreateSeraphimUnitEngineerBuildingEffects( self, unitBeingBuilt, self:GetBlueprint().General.BuildBones.BuildEffectBones, self.BuildEffectsBag )
     end,
 }
-TypeClass = PAB0104
+TypeClass = SAB0104