+++ mods/BrewLAN_Plenae/ExpertCamo/mod_info.lua
@@ -1,8 +1,8 @@
 name = "Expert Camo"
-uid = "2529ea71-93ef-41a6-b552-CAMMO000000001"
-version = 0.0
+uid = "2529ea71-93ef-41a6-b552-CAMMO000000010"
+version = 1
 copyright = "© 2017 Sean Wheeldon"
-description = "Some units have different textures on some maps."
+description = "Provides framework support for quickly generated reskins of units in mod folders 'TerrainMeshes' and 'CustomMeshes'."
 author = "Sean Wheeldon"
 --icon = "/mods/BrewLAN_Plenae/ExperimentalIcons/Experimental-Icon-Overhaul_Mod_Icon.dds"
 selectable = true